refactor(tools): move to audio package
Move tool.js to audio package.
